+++ common/datablocks/src/kernels/mod.rs
@@ -5,7 +5,7 @@
 #[cfg(test)]
 mod data_block_concat_test;
 #[cfg(test)]
-mod data_block_groupby_test;
+mod data_block_group_by_test;
 #[cfg(test)]
 mod data_block_scatter_test;
 #[cfg(test)]
@@ -14,7 +14,7 @@ mod data_block_sort_test;
 mod data_block_take_test;
 
 mod data_block_concat;
-mod data_block_groupby;
+mod data_block_group_by;
 mod data_block_scatter;
 mod data_block_sort;
 mod data_block_take;
