+++ common/planners/src/plan_aggregator_final.rs
@@ -14,7 +14,7 @@ pub struct AggregatorFinalPlan {
     pub aggr_expr: Vec<Expression>,
     pub group_expr: Vec<Expression>,
     pub schema: DataSchemaRef,
-    pub schema_before_groupby: DataSchemaRef,
+    pub schema_before_group_by: DataSchemaRef,
     pub input: Arc<PlanNode>,
 }
