+++ fusequery/query/src/pipelines/transforms/mod.rs
@@ -7,8 +7,8 @@ pub use transform_aggregator_partial::AggregatorPartialTransform;
 pub use transform_expression::ExpressionTransform;
 pub use transform_expression_executor::ExpressionExecutor;
 pub use transform_filter::FilterTransform;
-pub use transform_groupby_final::GroupByFinalTransform;
-pub use transform_groupby_partial::GroupByPartialTransform;
+pub use transform_group_by_final::GroupByFinalTransform;
+pub use transform_group_by_partial::GroupByPartialTransform;
 pub use transform_limit::LimitTransform;
 pub use transform_limit_by::LimitByTransform;
 pub use transform_projection::ProjectionTransform;
@@ -26,9 +26,9 @@ mod transform_expression_test;
 #[cfg(test)]
 mod transform_filter_test;
 #[cfg(test)]
-mod transform_groupby_final_test;
+mod transform_group_by_final_test;
 #[cfg(test)]
-mod transform_groupby_partial_test;
+mod transform_group_by_partial_test;
 #[cfg(test)]
 mod transform_limit_by_test;
 #[cfg(test)]
@@ -45,8 +45,8 @@ mod transform_aggregator_partial;
 mod transform_expression;
 mod transform_expression_executor;
 mod transform_filter;
-mod transform_groupby_final;
-mod transform_groupby_partial;
+mod transform_group_by_final;
+mod transform_group_by_partial;
 mod transform_limit;
 mod transform_limit_by;
 mod transform_projection;
