+++ fusequery/query/src/pipelines/transforms/transform_aggregator_final.rs
@@ -29,12 +29,12 @@ pub struct AggregatorFinalTransform {
 impl AggregatorFinalTransform {
     pub fn try_create(
         schema: DataSchemaRef,
-        schema_before_groupby: DataSchemaRef,
+        schema_before_group_by: DataSchemaRef,
         exprs: Vec<Expression>,
     ) -> Result<Self> {
         let funcs = exprs
             .iter()
-            .map(|expr| expr.to_aggregate_function(&schema_before_groupby))
+            .map(|expr| expr.to_aggregate_function(&schema_before_group_by))
             .collect::<Result<Vec<_>>>()?;
         Ok(AggregatorFinalTransform {
             funcs,
